Implement active request management and cancellation for MCP tools
- Added `ActiveRequests` struct to manage active MCP requests and their cancellation tokens.
- Implemented methods for registering, cancelling, and completing requests.
- Enhanced `spawn_execution` to support cancellation during tool execution.
- Refactored `run_mcp_server` to utilize `ActiveRequests` for handling incoming requests.
- Updated tool execution logic in `tools.rs` to support cancellation for `text_to_speech`.
- Removed unused `types.rs` file as its contents were integrated into other modules.
