+++ src/mcp/server.rs
@@ -1,99 +1,9 @@
 use anyhow::Result;
 use serde_json::Value;
-use std::fs;
 use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
 
-use crate::mcp::handlers;
-use crate::mcp::tools::get_tool_definitions;
-use crate::mcp::types::*;
-
-const MCP_VERSION: &str = "2025-03-26";
-const INSTRUCTIONS_ENV_VAR: &str = "VOICEVOX_MCP_INSTRUCTIONS";
-const INSTRUCTIONS_FILE: &str = "VOICEVOX.md";
-
-fn load_instructions() -> Option<String> {
-    use std::path::{Path, PathBuf};
-
-    fn try_load(path: &Path, description: &str) -> Option<String> {
-        eprintln!(
-            "Trying instructions from {}: {}",
-            description,
-            path.display()
-        );
-        match fs::read_to_string(path) {
-            Ok(content) => {
-                eprintln!("Loaded instructions from: {}", path.display());
-                Some(content)
-            }
-            Err(e) if e.kind() != std::io::ErrorKind::NotFound => {
-                eprintln!("Error loading instructions from {}: {}", path.display(), e);
-                None
-            }
-            _ => None,
-        }
-    }
-
-    // 1. Environment variable: VOICEVOX_MCP_INSTRUCTIONS (highest priority)
-    if let Ok(custom_path) = std::env::var(INSTRUCTIONS_ENV_VAR) {
-        let path = Path::new(&custom_path);
-        eprintln!(
-            "Trying instructions from environment variable: {}",
-            path.display()
-        );
-        match fs::read_to_string(path) {
-            Ok(content) => {
-                eprintln!("Loaded instructions from: {}", path.display());
-                return Some(content);
-            }
-            Err(e) => {
-                eprintln!("Could not load instructions from {}: {}", path.display(), e);
-            }
-        }
-    }
-
-    // 2. XDG user config: $XDG_CONFIG_HOME/voicevox/VOICEVOX.md (user-specific settings)
-    let xdg_config_var = std::env::var("XDG_CONFIG_HOME");
-    if let Ok(ref xdg_config) = xdg_config_var {
-        let path = PathBuf::from(xdg_config)
-            .join("voicevox")
-            .join(INSTRUCTIONS_FILE);
-        if let Some(content) = try_load(&path, "XDG_CONFIG_HOME") {
-            return Some(content);
-        }
-    }
-
-    // 3. Config fallback: ~/.config/voicevox/VOICEVOX.md (only when XDG_CONFIG_HOME is not set)
-    if xdg_config_var.is_err() {
-        if let Ok(home) = std::env::var("HOME") {
-            let path = PathBuf::from(home)
-                .join(".config")
-                .join("voicevox")
-                .join(INSTRUCTIONS_FILE);
-            if let Some(content) = try_load(&path, "~/.config") {
-                return Some(content);
-            }
-        }
-    }
-
-    // 4. Executable directory: VOICEVOX.md bundled with the binary (distribution default)
-    if let Ok(exe_path) = std::env::current_exe() {
-        if let Some(exe_dir) = exe_path.parent() {
-            let path = exe_dir.join(INSTRUCTIONS_FILE);
-            if let Some(content) = try_load(&path, "executable directory") {
-                return Some(content);
-            }
-        }
-    }
-
-    // 5. Current directory: VOICEVOX.md in working directory (development use)
-    let path = PathBuf::from(INSTRUCTIONS_FILE);
-    if let Some(content) = try_load(&path, "current directory") {
-        return Some(content);
-    }
-
-    eprintln!("No VOICEVOX.md found in any location");
-    None
-}
+use crate::mcp::protocol::{JsonRpcResponse, INVALID_REQUEST, PARSE_ERROR};
+use crate::mcp::requests::ActiveRequests;
 
 pub async fn run_mcp_server() -> Result<()> {
     let stdin = tokio::io::stdin();
@@ -101,57 +11,16 @@ pub async fn run_mcp_server() -> Result<()> {
     let reader = BufReader::new(stdin);
     let mut lines = reader.lines();
 
+    let active_requests = ActiveRequests::new();
     let mut shutdown = tokio::spawn(async {
         let _ = tokio::signal::ctrl_c().await;
     });
 
     loop {
         tokio::select! {
-            line = lines.next_line() => {
-                match line? {
-                    Some(line) if !line.trim().is_empty() => {
-                        let raw_request: Value = match serde_json::from_str(&line) {
-                            Ok(req) => req,
-                            Err(_) => {
-                                let id = serde_json::from_str::<Value>(&line)
-                                    .ok()
-                                    .and_then(|v| v.get("id").cloned())
-                                    .unwrap_or(Value::Number(serde_json::Number::from(0)));
-                                let error_response =
-                                    JsonRpcResponse::error(id, PARSE_ERROR, "Parse error".to_string());
-                                if let Ok(response_str) = serde_json::to_string(&error_response) {
-                                    let _ = stdout.write_all(response_str.as_bytes()).await;
-                                    let _ = stdout.write_all(b"\n").await;
-                                    let _ = stdout.flush().await;
-                                }
-                                continue;
-                            }
-                        };
-
-                        if raw_request.get("method").is_some() {
-                            if let Some(response) = handle_request(raw_request).await {
-                                if let Ok(response_str) = serde_json::to_string(&response) {
-                                    let _ = stdout.write_all(response_str.as_bytes()).await;
-                                    let _ = stdout.write_all(b"\n").await;
-                                    let _ = stdout.flush().await;
-                                }
-                            }
-                        } else {
-                            let id = raw_request
-                                .get("id")
-                                .cloned()
-                                .unwrap_or(Value::Number(serde_json::Number::from(0)));
-                            let response =
-                                JsonRpcResponse::error(id, INVALID_REQUEST, "Invalid request".to_string());
-                            if let Ok(response_str) = serde_json::to_string(&response) {
-                                let _ = stdout.write_all(response_str.as_bytes()).await;
-                                let _ = stdout.write_all(b"\n").await;
-                                let _ = stdout.flush().await;
-                            }
-                        }
-                    }
-                    None => break,
-                    _ => continue,
+            line_result = lines.next_line() => {
+                if !process_line(line_result?, &active_requests, &mut stdout).await {
+                    break;
                 }
             }
             _ = &mut shutdown => break,
@@ -161,154 +30,80 @@ pub async fn run_mcp_server() -> Result<()> {
     Ok(())
 }
 
-async fn handle_request(request: Value) -> Option<JsonRpcResponse> {
-    let id = request.get("id").cloned();
-    let method = request.get("method").and_then(|v| v.as_str()).unwrap_or("");
+async fn process_line(
+    line_option: Option<String>,
+    active_requests: &ActiveRequests,
+    stdout: &mut tokio::io::Stdout,
+) -> bool {
+    let line = match line_option {
+        Some(line) if !line.trim().is_empty() => line,
+        Some(_) => return true, // Empty line, continue
+        None => return false,   // EOF, terminate
+    };
 
-    match method {
-        "initialize" => {
-            let id = id.unwrap_or(Value::Number(serde_json::Number::from(0)));
-            let result = InitializeResult {
-                protocol_version: MCP_VERSION.to_string(),
-                server_info: ServerInfo {
-                    name: "voicevox-mcp".to_string(),
-                    version: env!("CARGO_PKG_VERSION").to_string(),
-                },
-                capabilities: ServerCapabilities {
-                    tools: serde_json::Map::new(),
-                },
-                instructions: load_instructions(),
-            };
+    let raw_request = match parse_json_request(&line, stdout).await {
+        Some(request) => request,
+        None => return true, // Parse error handled, continue
+    };
 
-            match serde_json::to_value(result) {
-                Ok(value) => Some(JsonRpcResponse::success(id, value)),
-                Err(_) => Some(JsonRpcResponse::error(
-                    id,
-                    INTERNAL_ERROR,
-                    "Failed to serialize response".to_string(),
-                )),
-            }
-        }
-        "notifications/initialized" => None,
-        "tools/list" => {
-            let id = id.unwrap_or(Value::Number(serde_json::Number::from(0)));
-            let result = ToolsListResult {
-                tools: get_tool_definitions(),
-            };
-            match serde_json::to_value(result) {
-                Ok(value) => Some(JsonRpcResponse::success(id, value)),
-                Err(_) => Some(JsonRpcResponse::error(
-                    id,
-                    INTERNAL_ERROR,
-                    "Failed to serialize response".to_string(),
-                )),
-            }
-        }
-        "tools/call" => {
-            let id = id.unwrap_or(Value::Number(serde_json::Number::from(0)));
-            if let Some(params) = request.get("params") {
-                if let Some(params_obj) = params.as_object() {
-                    let tool_name = params_obj
-                        .get("name")
-                        .and_then(|v| v.as_str())
-                        .unwrap_or("");
+    if raw_request.get("method").is_some() {
+        handle_message(raw_request, active_requests, stdout).await;
+    } else {
+        send_invalid_request_error(&raw_request, stdout).await;
+    }
 
-                    let arguments = params_obj
-                        .get("arguments")
-                        .cloned()
-                        .unwrap_or(Value::Object(serde_json::Map::new()));
+    true
+}
 
-                    match tool_name {
-                        "text_to_speech" => {
-                            match handlers::handle_text_to_speech(arguments).await {
-                                Ok(result) => match serde_json::to_value(result) {
-                                    Ok(value) => Some(JsonRpcResponse::success(id.clone(), value)),
-                                    Err(_) => Some(JsonRpcResponse::error(
-                                        id.clone(),
-                                        INTERNAL_ERROR,
-                                        "Failed to serialize response".to_string(),
-                                    )),
-                                },
-                                Err(e) => {
-                                    let error_result = ToolCallResult {
-                                        content: vec![ToolContent {
-                                            content_type: "text".to_string(),
-                                            text: format!("Synthesis error: {e}"),
-                                        }],
-                                        is_error: Some(true),
-                                    };
-                                    match serde_json::to_value(error_result) {
-                                        Ok(value) => {
-                                            Some(JsonRpcResponse::success(id.clone(), value))
-                                        }
-                                        Err(_) => Some(JsonRpcResponse::error(
-                                            id.clone(),
-                                            INTERNAL_ERROR,
-                                            "Failed to serialize error response".to_string(),
-                                        )),
-                                    }
-                                }
-                            }
-                        }
-                        "list_voice_styles" => {
-                            match handlers::handle_list_voice_styles(arguments).await {
-                                Ok(result) => match serde_json::to_value(result) {
-                                    Ok(value) => Some(JsonRpcResponse::success(id.clone(), value)),
-                                    Err(_) => Some(JsonRpcResponse::error(
-                                        id.clone(),
-                                        INTERNAL_ERROR,
-                                        "Failed to serialize response".to_string(),
-                                    )),
-                                },
-                                Err(e) => {
-                                    let error_result = ToolCallResult {
-                                        content: vec![ToolContent {
-                                            content_type: "text".to_string(),
-                                            text: format!("Error getting voices: {e}"),
-                                        }],
-                                        is_error: Some(true),
-                                    };
-                                    match serde_json::to_value(error_result) {
-                                        Ok(value) => {
-                                            Some(JsonRpcResponse::success(id.clone(), value))
-                                        }
-                                        Err(_) => Some(JsonRpcResponse::error(
-                                            id.clone(),
-                                            INTERNAL_ERROR,
-                                            "Failed to serialize error response".to_string(),
-                                        )),
-                                    }
-                                }
-                            }
-                        }
-                        _ => Some(JsonRpcResponse::error(
-                            id.clone(),
-                            METHOD_NOT_FOUND,
-                            format!("Unknown tool: {tool_name}"),
-                        )),
-                    }
-                } else {
-                    Some(JsonRpcResponse::error(
-                        id.clone(),
-                        INVALID_PARAMS,
-                        "Invalid params".to_string(),
-                    ))
-                }
-            } else {
-                Some(JsonRpcResponse::error(
-                    id.clone(),
-                    INVALID_PARAMS,
-                    "Missing params".to_string(),
-                ))
-            }
+async fn parse_json_request(line: &str, stdout: &mut tokio::io::Stdout) -> Option<Value> {
+    match serde_json::from_str(line) {
+        Ok(request) => Some(request),
+        Err(_) => {
+            let id = extract_id_from_invalid_json(line);
+            let error_response = JsonRpcResponse::error(id, PARSE_ERROR, "Parse error".to_string());
+            send_response(&error_response, stdout).await;
+            None
         }
-        _ => {
-            let id = id.unwrap_or(Value::Number(serde_json::Number::from(0)));
-            Some(JsonRpcResponse::error(
-                id,
-                METHOD_NOT_FOUND,
-                format!("Method not found: {method}"),
-            ))
-        }
+    }
+}
+
+fn extract_id_from_invalid_json(line: &str) -> Value {
+    serde_json::from_str::<Value>(line)
+        .ok()
+        .and_then(|v| v.get("id").cloned())
+        .unwrap_or(Value::Number(serde_json::Number::from(0)))
+}
+
+async fn send_invalid_request_error(raw_request: &Value, stdout: &mut tokio::io::Stdout) {
+    let id = raw_request
+        .get("id")
+        .cloned()
+        .unwrap_or(Value::Number(serde_json::Number::from(0)));
+    let response = JsonRpcResponse::error(id, INVALID_REQUEST, "Invalid request".to_string());
+    send_response(&response, stdout).await;
+}
+
+async fn send_response(response: &JsonRpcResponse, stdout: &mut tokio::io::Stdout) {
+    if let Ok(response_str) = serde_json::to_string(response) {
+        let _ = stdout.write_all(response_str.as_bytes()).await;
+        let _ = stdout.write_all(b"\n").await;
+        let _ = stdout.flush().await;
+    }
+}
+
+async fn handle_message(
+    request: Value,
+    active_requests: &ActiveRequests,
+    stdout: &mut tokio::io::Stdout,
+) {
+    // Handle notifications (no response expected)
+    if request.get("id").is_none() {
+        crate::mcp::protocol::handle_notification(request, active_requests).await;
+        return;
+    }
+
+    // Handle requests (response expected)
+    if let Some(response) = crate::mcp::protocol::process_request(request, active_requests).await {
+        send_response(&response, stdout).await;
     }
 }
